refactor: postprocessing move to unit test
